backend/requirements.txt
flask
tensorflow
pillow
numpy
python-dotenv
requests
flask-cors
gunicorn
